:recycle: Create common base class for simplification steps
